+++ HISTORY.md
@@ -130,3 +130,4 @@
  **version 1.2.7**
   - Bug fix in the verbose output of the spectral_extraction module
   - Added capability for JWST/NIRSPEC data
+  - Moved spectral and time rebinning for spectra from instruments to TSO module.